# docs/swagger.yaml
basePath: /api/v1
definitions:
  types.FailureErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  types.FailureResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  types.GetMeResponse:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
  types.LoginRequest:
    properties:
      identity:
        type: string
      password:
        type: string
    type: object
  types.LoginSuccessData:
    properties:
      token:
        type: string
    type: object
  types.LoginSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/types.LoginSuccessData'
      status:
        type: string
    type: object
  types.RegisterRequest:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  types.SuccessResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  types.TaskResponse:
    properties:
      assignee_id:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      status:
        type: string
      team_id:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: This is an API of auth-service
  title: Service task  Swagger
  version: "1.0"
paths:
  /auth/get-me:
    get:
      description: Get current user ID and email
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetMeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get current user info
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: username or email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.LoginSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/refresh:
    post:
      description: Refresh jwt token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Refresh jwt token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register
      parameters:
      - description: username or email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.LoginSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      summary: Register
      tags:
      - auth
  /system/invalid-assignees:
    get:
      description: '"Найти задачи, где assignee не является членом команды этой задачи".'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TaskResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: getTasksWithInvalidAssignee оборачивает запрос info
      tags:
      - complex-sql
  /system/teams-stats:
    get:
      description: '"Получить для каждой команды: название, количество участников,
        количество задач в статусе done за последние 7 дней".'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TaskResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: getTeamStats оборачивает сложный SQL info
      tags:
      - complex-sql
  /system/teams-top-creators:
    get:
      description: '"Получить топ-3 пользователя по количеству созданных задач в каждой
        команде за месяц".'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TaskResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: getTopCreatorsByTeam оборачивает сложный SQL с оконной функцией info
      tags:
      - complex-sql
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"